refactor into components finished
